feat: wire adding as one undo operation
Previously clicking on a handle would add a new wire to the graph as a command. Then moving and finally releasing it would issue a second command to change its position. This would result in 2 undo entries, that would need to be undone separately. However, for the user this is one action.

This commit makes it so that a new wire is temporarily added to the DOM after a handle is clicked. Once the wire is released it is removed one AddWireCommand is issued. This command adds a new entry to the graph which is then dynamically rendered. This allows the use to undo the action with one click.